Minor tweaks
commented out a call to timeInit(), which resets the RTC to the last sketch compile time on power restore. This was handy for testing purposes, but isn't necessary when the RTC has a battery and can keep it's own time.
Also, fixed a bug where the year would show up as "20 0' when setting it and the decade was 0.